requirements.txt
streamlit
pandas
gspread
oauth2client
langchain
langchain-community
langchain-google-genai
chromadb
google-generativeai
python-dotenv